moving rings and asteroids out, something broke
